Updated VIPER template to the format @ed and @jason agreed on
