+++ Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___Interactor.swift
@@ -6,16 +6,20 @@
 //  ___COPYRIGHT___
 //
 
-import Foundation
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___InteractorInput: AnyObject {
+}
+
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___InteractorOutput: AnyObject {
+}
 
 final class ___VARIABLE_moduleName___Interactor {
 
     weak var output: ___VARIABLE_moduleName___InteractorOutput?
-    
 }
 
 // MARK: - ___VARIABLE_moduleName___InteractorInput
 
 extension ___VARIABLE_moduleName___Interactor: ___VARIABLE_moduleName___InteractorInput {
-
 }
