--- Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___InteractorInput.swift
@@ -1,14 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import Foundation
-
-___VARIABLE_sourceryAnnotationForProtocols___
-protocol ___VARIABLE_moduleName___InteractorInput: AnyObject {
-
-}
