+++ Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___Module.swift
@@ -8,32 +8,40 @@
 
 import UIKit
 
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___Input: AnyObject {
+}
+
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___Output: AnyObject {
+}
+
 final class ___VARIABLE_moduleName___Module {
-    
+
     let view: ___VARIABLE_moduleName___ViewController
     let presenter: ___VARIABLE_moduleName___Presenter
     let router: ___VARIABLE_moduleName___Router
     let interactor: ___VARIABLE_moduleName___Interactor
 
-    var output: ___VARIABLE_moduleName___Output? { 
+    var output: ___VARIABLE_moduleName___Output? {
         get { return presenter.output }
         set { presenter.output = newValue }
     }
 
-    var input: ___VARIABLE_moduleName___Input? { 
+    var input: ___VARIABLE_moduleName___Input? {
         return presenter
     }
 
-    init() { 
+    init() {
         self.view = ___VARIABLE_moduleName___ViewController()
         self.presenter = ___VARIABLE_moduleName___Presenter()
         self.router = ___VARIABLE_moduleName___Router()
         self.interactor = ___VARIABLE_moduleName___Interactor()
 
         view.output = presenter
-        
+
         presenter.view = view
-        presenter.router = router 
+        presenter.router = router
         presenter.interactor = interactor
         presenter.output = output
 
@@ -41,5 +49,4 @@ final class ___VARIABLE_moduleName___Module {
 
         router.view = view
     }
-
 }
