+++ Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___Presenter.swift
@@ -6,8 +6,6 @@
 //  ___COPYRIGHT___
 //
 
-import Foundation
-
 final class ___VARIABLE_moduleName___Presenter {
 
     weak var view: ___VARIABLE_moduleName___ViewInput?
@@ -15,27 +13,23 @@ final class ___VARIABLE_moduleName___Presenter {
     var interactor: ___VARIABLE_moduleName___InteractorInput?
 
     var output: ___VARIABLE_moduleName___Output?
-
 }
 
 // MARK: - ___VARIABLE_moduleName___ViewOutput
 
 extension ___VARIABLE_moduleName___Presenter: ___VARIABLE_moduleName___ViewOutput {
 
-    func viewDidLoad() { 
+    func viewDidLoad() {
         view?.configure()
     }
-    
 }
 
 // MARK: - ___VARIABLE_moduleName___InteractorOutput
 
-extension ___VARIABLE_moduleName___Presenter: ___VARIABLE_moduleName___InteractorOutput { 
-
+extension ___VARIABLE_moduleName___Presenter: ___VARIABLE_moduleName___InteractorOutput {
 }
 
 // MARK: - ___VARIABLE_moduleName___Input
 
-extension ___VARIABLE_moduleName___Presenter: ___VARIABLE_moduleName___Input { 
-    
+extension ___VARIABLE_moduleName___Presenter: ___VARIABLE_moduleName___Input {
 }
