+++ Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___Router.swift
@@ -8,18 +8,20 @@
 
 import UIKit
 
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___RouterInput: AnyObject {
+}
+
 final class ___VARIABLE_moduleName___Router {
-    
+
     weak var view: ___VARIABLE_moduleName___ViewInput?
 
-    private var viewController: UIViewController? { 
+    private var viewController: UIViewController? {
         return view as? UIViewController
     }
-
 }
 
 // MARK: - ___VARIABLE_moduleName___RouterInput
 
 extension ___VARIABLE_moduleName___Router: ___VARIABLE_moduleName___RouterInput {
-
 }
