+++ Templates/VIPER Module.xctemplate/___VARIABLE_moduleName___ViewController.swift
@@ -8,23 +8,32 @@
 
 import UIKit
 
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___ViewInput: AnyObject {
+
+    func configure()
+}
+
+___VARIABLE_sourceryAnnotationForProtocols___
+protocol ___VARIABLE_moduleName___ViewOutput: AnyObject {
+
+    func viewDidLoad()
+}
+
 final class ___VARIABLE_moduleName___ViewController: UIViewController {
-    
+
     var output: ___VARIABLE_moduleName___ViewOutput?
 
-    override func viewDidLoad() { 
+    override func viewDidLoad() {
         super.viewDidLoad()
         output?.viewDidLoad()
     }
-
 }
 
 // MARK: - ___VARIABLE_moduleName___ViewInput
 
-extension ___VARIABLE_moduleName___ViewController: ___VARIABLE_moduleName___ViewInput { 
+extension ___VARIABLE_moduleName___ViewController: ___VARIABLE_moduleName___ViewInput {
 
-    func configure() { 
-        
+    func configure() {
     }
-
 }
